+++ hnix-store-json/src/System/Nix/JSON.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE MonoLocalBinds #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-|
 Description : JSON serialization
@@ -15,40 +16,38 @@ module System.Nix.JSON
 import Control.Applicative ((<|>))
 import Crypto.Hash (Digest)
 import Data.Aeson
+import Data.Aeson.Key qualified
 import Data.Aeson.KeyMap qualified
 import Data.Aeson.Types (Parser)
 import Data.Aeson.Types qualified
 import Data.Attoparsec.Text qualified
-import Data.Char qualified
-import Data.Constraint.Extras (Has(has))
 import Data.Dependent.Sum
 import Data.Foldable (toList)
+import Data.Map.Monoidal qualified
 import Data.Map.Strict qualified
-import Data.Maybe (maybeToList)
+import Data.Maybe (fromMaybe, maybeToList)
 import Data.Set qualified
 import Data.Some
 import Data.Text (Text)
 import Data.Text qualified
-import Data.Text.Lazy qualified
-import Data.Text.Lazy.Builder qualified
+import Data.These
+import Data.These.Combinators
 import Data.Time (diffUTCTime)
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
-import Deriving.Aeson
-
-import System.Nix.Base (baseEncodingToText, textToBaseEncoding)
 import System.Nix.Base qualified
 import System.Nix.Build (BuildResult(..), BuildSuccess(..), BuildFailure(..), BuildSuccessStatus(..), BuildFailureStatus(..))
 import System.Nix.ContentAddress
+import System.Nix.Derivation
 import System.Nix.DerivedPath (DerivedPath(..), OutputsSpec(..), SingleDerivedPath(..))
 import System.Nix.Hash
-import System.Nix.OutputName (OutputName)
+import System.Nix.OutputName (OutputName, mkOutputName)
 import System.Nix.OutputName qualified
-import System.Nix.Realisation (BuildTraceKey(..), Realisation, RealisationWithId(..))
-import System.Nix.Realisation qualified
-import System.Nix.Signature (Signature)
+import System.Nix.Realisation (BuildTraceKey(..), Realisation(..), RealisationWithId(..))
+import System.Nix.Signature (Signature, NamedSignature(..))
 import System.Nix.Signature qualified
 import System.Nix.StorePath (StorePath, StorePathName, StorePathHashPart, mkStorePathName, unStorePathName, parseBasePathFromText)
 import System.Nix.StorePath qualified
+import System.Nix.StorePath.Metadata (Metadata(..), StorePathTrust(..))
 
 instance ToJSON StorePathName where
   toJSON = toJSON . System.Nix.StorePath.unStorePathName
@@ -93,6 +92,109 @@ instance FromJSONKey StorePath where
   fromJSONKey = FromJSONKeyTextParser $
     either (fail . show) pure . parseBasePathFromText
 
+deriving newtype instance FromJSON DerivedPathMap
+deriving newtype instance ToJSON DerivedPathMap
+
+instance FromJSON ChildNode where
+  parseJSON = withObject "ChildNode" $ \obj -> do
+    outputs <- obj .: "outputs"
+    dynamicOutputs <- obj .: "dynamicOutputs"
+    ChildNode <$> case (Data.Set.null outputs, Data.Map.Monoidal.null dynamicOutputs) of
+      (False, True) -> pure $ This outputs
+      (True, False) -> pure $ That dynamicOutputs
+      (False, False) -> pure $ These outputs dynamicOutputs
+      (True, True) -> fail "outputs and dynamic outputs cannot both be empty"
+
+instance ToJSON ChildNode where
+  toJSON (ChildNode cn) = object
+    [ "outputs" .= fromMaybe Data.Set.empty (justHere cn)
+    , "dynamicOutputs" .= fromMaybe Data.Map.Monoidal.empty (justThere cn)
+    ]
+
+-- | Input-addressed derivation output JSON instance
+instance FromJSON InputAddressedDerivationOutput where
+  parseJSON = withObject "InputAddressedDerivationOutput" $ \obj ->
+    InputAddressedDerivationOutput <$> obj .: "path"
+
+instance ToJSON InputAddressedDerivationOutput where
+  toJSON (InputAddressedDerivationOutput path) =
+    object ["path" .= path]
+
+-- | Fixed derivation output JSON instance
+instance FromJSON FixedDerivationOutput where
+  parseJSON = withObject "FixedDerivationOutput" $ \obj -> do
+    method <- obj .: "method" >>= either fail pure . textToMethod
+    HashJSON hash <- obj .: "hash"
+    pure $ FixedDerivationOutput method hash
+
+instance ToJSON FixedDerivationOutput where
+  toJSON (FixedDerivationOutput method hash) =
+    object
+      [ "method" .= methodToText method
+      , "hash" .= HashJSON hash
+      ]
+
+-- | Content-addressed derivation output JSON instance
+instance FromJSON ContentAddressedDerivationOutput where
+  parseJSON = withObject "ContentAddressedDerivationOutput" $ \obj -> do
+    method <- obj .: "method" >>= either fail pure . textToMethod
+    hashAlgo <- obj .: "hashAlgo" >>= either fail pure . textToAlgo
+    pure $ ContentAddressedDerivationOutput method hashAlgo
+
+instance ToJSON ContentAddressedDerivationOutput where
+  toJSON (ContentAddressedDerivationOutput method (Some hashAlgo)) =
+    object
+      [ "method" .= methodToText method
+      , "hashAlgo" .= algoToText hashAlgo
+      ]
+
+-- Helper to parse DerivationOutputs from JSON
+parseDerivationOutputs :: Object -> Parser DerivationOutputs
+parseDerivationOutputs obj = do
+  let outputPairs = Data.Aeson.KeyMap.toList obj
+  case outputPairs of
+    [] -> pure $ DerivationType_InputAddressing :=> mempty
+    (_, firstVal) : _ -> do
+      -- Parse the first output to determine the derivation type
+      withObject "first output" (\first -> do
+        -- Determine type by checking which fields are present
+        let hasPath = Data.Aeson.KeyMap.member "path" first
+            hasMethod = Data.Aeson.KeyMap.member "method" first
+            hasHash = Data.Aeson.KeyMap.member "hash" first
+            hasHashAlgo = Data.Aeson.KeyMap.member "hashAlgo" first
+
+        case (hasPath, hasMethod, hasHash, hasHashAlgo) of
+          (True, False, False, False) -> do
+            -- Input-addressed
+            outputs <- Data.Map.Strict.fromList <$> mapM (\(k, v) -> do
+              outputName <- either (fail . show) pure $ mkOutputName $ Data.Aeson.Key.toText k
+              output <- parseJSON v
+              pure (outputName, output)) outputPairs
+            pure $ DerivationType_InputAddressing :=> outputs
+          (False, True, True, False) -> do
+            -- Fixed
+            outputs <- Data.Map.Strict.fromList <$> mapM (\(k, v) -> do
+              outputName <- either (fail . show) pure $ mkOutputName $ Data.Aeson.Key.toText k
+              output <- parseJSON v
+              pure (outputName, output)) outputPairs
+            pure $ DerivationType_Fixed :=> outputs
+          (False, True, False, True) -> do
+            -- Content-addressed
+            outputs <- Data.Map.Strict.fromList <$> mapM (\(k, v) -> do
+              outputName <- either (fail . show) pure $ mkOutputName $ Data.Aeson.Key.toText k
+              output <- parseJSON v
+              pure (outputName, output)) outputPairs
+            pure $ DerivationType_ContentAddressing :=> outputs
+          _ -> fail $ "Invalid output format: " <> show (hasPath, hasMethod, hasHash, hasHashAlgo)
+        ) firstVal
+
+-- Helper to serialize DerivationOutputs to JSON
+derivationOutputsToJSON :: DerivationOutputs -> Value
+derivationOutputsToJSON (ty :=> outputs) = case ty of
+  DerivationType_InputAddressing -> toJSON outputs
+  DerivationType_Fixed -> toJSON outputs
+  DerivationType_ContentAddressing -> toJSON outputs
+
 instance FromJSONKey StorePathName where
   fromJSONKey = FromJSONKeyTextParser $ either (fail . show) pure . mkStorePathName
 
@@ -104,48 +206,48 @@ deriving newtype instance ToJSON OutputName
 deriving newtype instance FromJSONKey OutputName
 deriving newtype instance ToJSONKey OutputName
 
-instance ToJSON (BuildTraceKey OutputName) where
-  toJSON =
-    toJSON
-    . Data.Text.Lazy.toStrict
-    . Data.Text.Lazy.Builder.toLazyText
-    . System.Nix.Realisation.buildTraceKeyBuilder
-        System.Nix.OutputName.outputNameToText
+-- | TODO: hacky, we need to stop assuming StoreDir for
+-- StorePath to and from JSON
+instance FromJSON Derivation where
+  parseJSON = withObject "Derivation" $ \v -> do
+    name <- v .: "name"
+    inputs <- v .: "inputs" >>= \inputsObj -> DerivationInputs
+      <$> inputsObj .: "srcs"
+      <*> inputsObj .: "drvs"
+    outputs <- v .: "outputs" >>= parseDerivationOutputs
+    Derivation name outputs
+      <$> pure inputs
+      <*> v .: "system"
+      <*> v .: "builder"
+      <*> v .: "args"
+      <*> v .: "env"
 
-  toEncoding =
-    toEncoding
-    . Data.Text.Lazy.toStrict
-    . Data.Text.Lazy.Builder.toLazyText
-    . System.Nix.Realisation.buildTraceKeyBuilder
-        System.Nix.OutputName.outputNameToText
+instance ToJSON Derivation where
+  toJSON (Derivation name outputs (DerivationInputs inputSrcs inputDrvs) system builder args env) =
+    object [ "name" .= name
+           , "outputs" .= derivationOutputsToJSON outputs
+           , "inputs" .= object
+              [ "srcs" .= inputSrcs
+              , "drvs" .= inputDrvs
+              ]
+           , "system" .= system
+           , "builder" .= builder
+           , "args" .= args
+           , "env" .= env
+           ]
 
-instance ToJSONKey (BuildTraceKey OutputName) where
-  toJSONKey =
-    Data.Aeson.Types.toJSONKeyText
-    $ Data.Text.Lazy.toStrict
-    . Data.Text.Lazy.Builder.toLazyText
-    . System.Nix.Realisation.buildTraceKeyBuilder
-        System.Nix.OutputName.outputNameToText
+instance ToJSON BuildTraceKey where
+  toJSON btk =
+    object
+      [ "drvPath" .= buildTraceKeyDrvPath btk
+      , "outputName" .= buildTraceKeyOutput btk
+      ]
 
-instance FromJSON (BuildTraceKey OutputName) where
-  parseJSON =
-    withText "BuildTraceKey OutputName"
-    ( either
-        (fail . show)
-        pure
-    . System.Nix.Realisation.buildTraceKeyParser
-        System.Nix.OutputName.mkOutputName
-    )
-
-instance FromJSONKey (BuildTraceKey OutputName) where
-  fromJSONKey =
-    FromJSONKeyTextParser
-    ( either
-        (fail . show)
-        pure
-    . System.Nix.Realisation.buildTraceKeyParser
-        System.Nix.OutputName.mkOutputName
-    )
+instance FromJSON BuildTraceKey where
+  parseJSON = withObject "BuildTraceKey" $ \o ->
+    BuildTraceKey
+      <$> o .: "drvPath"
+      <*> o .: "outputName"
 
 instance ToJSON Signature where
   toJSON = toJSON . System.Nix.Signature.signatureToText
@@ -161,31 +263,37 @@ instance FromJSON Signature where
         System.Nix.Signature.signatureParser
     )
 
+instance ToJSON NamedSignature where
+  toJSON ns =
+    object
+      [ "keyName" .= publicKey ns
+      , "sig" .= sig ns
+      ]
+
+instance FromJSON NamedSignature where
+  parseJSON = withObject "NamedSignature" $ \o -> do
+    keyName <- o .: "keyName"
+    sigVal <- o .: "sig"
+    pure $ NamedSignature { publicKey = keyName, sig = sigVal }
+
 -- | Needed to avoid overlapping instances
 newtype HashJSON = HashJSON { unHashJSON :: DSum HashAlgo Digest }
   deriving (Eq, Show)
 
 instance ToJSON HashJSON where
   toJSON (HashJSON (algo :=> digest)) =
-    object
-      [ "algorithm" .= algoToText algo
-      -- We can parse others, but always render using base64
-      , "format" .= baseEncodingToText Base64
-      , "hash" .= encodeDigestWith Base64 digest
-      ]
+    -- SRI format: "algo-base64hash"
+    toJSON $ algoToText algo <> "-" <> encodeDigestWith Base64 digest
 
 instance FromJSON HashJSON where
-  parseJSON = withObject "HashJSON" $ \obj -> do
-    algoText <- obj .: "algorithm"
-    formatText <- obj .: "format"
-    hashText <- obj .: "hash"
-
-    Some algo <- either fail pure $ textToAlgo algoText
-    format <- either fail pure $ textToBaseEncoding formatText
-
-    digest <- has @NamedAlgo algo $ either fail pure $ decodeDigestWith format hashText
-
-    pure $ HashJSON (algo :=> digest)
+  parseJSON = withText "HashJSON" $ \t -> do
+    -- SRI format: "algo-base64hash"
+    let (algoText, rest) = Data.Text.breakOn "-" t
+    case Data.Text.uncons rest of
+      Nothing -> fail "HashJSON: missing '-' separator"
+      Just ('-', _hashText) ->
+        either fail (pure . HashJSON) $ mkNamedDigest algoText t
+      _ -> fail "HashJSON: missing '-' separator"
 
 instance ToJSON ContentAddress where
   toJSON (ContentAddress method digest) =
@@ -250,46 +358,35 @@ instance FromJSON DerivedPath where
           <$> obj .: "drvPath"
           <*> obj .: "outputs"
 
-data LowerLeading
-instance StringModifier LowerLeading where
-  getStringModifier "" = ""
-  getStringModifier (c:xs) = Data.Char.toLower c : xs
+instance ToJSON Realisation where
+  toJSON r =
+    object
+      [ "outPath" .= realisationOutPath r
+      , "signatures" .= realisationSignatures r
+      ]
 
-deriving
-  via CustomJSON
-    '[FieldLabelModifier
-       '[ StripPrefix "realisation"
-        , LowerLeading
-        , Rename "dependencies" "dependentRealisations"
-        ]
-     ] Realisation
-  instance ToJSON Realisation
-deriving
-  via CustomJSON
-    '[FieldLabelModifier
-       '[ StripPrefix "realisation"
-        , LowerLeading
-        , Rename "dependencies" "dependentRealisations"
-        ]
-     ] Realisation
-  instance FromJSON Realisation
+instance FromJSON Realisation where
+  parseJSON = withObject "Realisation" $ \o ->
+    Realisation
+      <$> o .: "outPath"
+      <*> o .:? "signatures" .!= mempty
 
 -- For a keyed version of Realisation
--- we use RealisationWithId (BuildTraceKey OutputName, Realisation)
--- instead of Realisation.id :: (BuildTraceKey OutputName)
+-- we use RealisationWithId (BuildTraceKey, Realisation)
+-- instead of Realisation.id :: BuildTraceKey
 -- field.
 instance ToJSON RealisationWithId where
   toJSON (RealisationWithId (drvOut, r)) =
-    case toJSON r of
-      Object o -> Object $ Data.Aeson.KeyMap.insert "id" (toJSON drvOut) o
-      _ -> error "absurd"
+    object
+      [ "key" .= drvOut
+      , "value" .= r
+      ]
 
 instance FromJSON RealisationWithId where
-  parseJSON v@(Object o) = do
-    r <- parseJSON @Realisation v
-    drvOut <- o .: "id"
+  parseJSON = withObject "RealisationWithId" $ \o -> do
+    drvOut <- o .: "key"
+    r <- o .: "value"
     pure (RealisationWithId (drvOut, r))
-  parseJSON x = fail $ "Expected Object but got " ++ show x
 
 -- BuildSuccessStatus enum to/from JSON as strings
 instance ToJSON BuildSuccessStatus where
@@ -342,15 +439,10 @@ instance ToJSON BuildResult where
   toJSON (BuildResult status timesBuilt startTime stopTime cpuUser cpuSystem) =
     case status of
       Right (BuildSuccess successStatus builtOutputs) ->
-        -- Convert Map (BuildTraceKey OutputName) Realisation to Map OutputName RealisationWithId
-        let builtOutputsForJSON = Data.Map.Strict.fromList
-              [ (outName, RealisationWithId (btk, r))
-              | (btk@(BuildTraceKey _hash outName), r) <- Data.Map.Strict.toList builtOutputs
-              ]
-        in object $ concat
+        object $ concat
           [ [ "success" .= True
             , "status" .= successStatus
-            , "builtOutputs" .= builtOutputsForJSON
+            , "builtOutputs" .= builtOutputs
             , "timesBuilt" .= timesBuilt
             , "startTime" .= (floor (realToFrac (diffUTCTime startTime (posixSecondsToUTCTime 0)) :: Double) :: Integer)
             , "stopTime" .= (floor (realToFrac (diffUTCTime stopTime (posixSecondsToUTCTime 0)) :: Double) :: Integer)
@@ -381,12 +473,7 @@ instance FromJSON BuildResult where
     buildResultStatus <- if success
       then do
         successStatus <- obj .: "status"
-        -- Parse as Map OutputName RealisationWithId, then convert to Map (BuildTraceKey OutputName) Realisation
-        builtOutputsWithId <- obj .: "builtOutputs" :: Parser (Data.Map.Strict.Map OutputName RealisationWithId)
-        let builtOutputs = Data.Map.Strict.fromList
-              [ (btk, r)
-              | (_outName, RealisationWithId (btk, r)) <- Data.Map.Strict.toList builtOutputsWithId
-              ]
+        builtOutputs <- obj .:? "builtOutputs" .!= mempty
         pure $ Right (BuildSuccess successStatus builtOutputs)
       else do
         failureStatus <- obj .: "status"
@@ -404,4 +491,45 @@ instance FromJSON BuildResult where
       , buildResultStopTime = stopTime
       , buildResultCpuUser = buildResultCpuUser
       , buildResultCpuSystem = buildResultCpuSystem
+      }
+
+-- | Metadata (path-info) JSON, version 3 format
+instance ToJSON (Metadata StorePath) where
+  toJSON m = object
+    [ "version" .= (3 :: Int)
+    , "storeDir" .= ("/nix/store" :: Text)
+    , "narHash" .= HashJSON (metadataNarHash m)
+    , "narSize" .= fromMaybe 0 (metadataNarBytes m)
+    , "references" .= metadataReferences m
+    , "ca" .= metadataContentAddress m
+    , "deriver" .= metadataDeriverPath m
+    , "registrationTime" .=
+        let t = metadataRegistrationTime m
+        in if t == posixSecondsToUTCTime 0
+           then Nothing @Integer
+           else Just (floor (realToFrac (diffUTCTime t (posixSecondsToUTCTime 0)) :: Double) :: Integer)
+    , "signatures" .= metadataSigs m
+    , "ultimate" .= (metadataTrust m == BuiltLocally)
+    ]
+
+instance FromJSON (Metadata StorePath) where
+  parseJSON = withObject "Metadata" $ \o -> do
+    HashJSON narHash <- o .: "narHash"
+    narSize <- o .: "narSize"
+    references <- o .:? "references" .!= mempty
+    ca <- o .:? "ca"
+    deriver <- o .:? "deriver"
+    regTime <- o .:? "registrationTime"
+    sigs <- o .:? "signatures" .!= mempty
+    ultimate <- o .:? "ultimate" .!= False
+    pure Metadata
+      { metadataDeriverPath = deriver
+      , metadataNarHash = narHash
+      , metadataReferences = references
+      , metadataRegistrationTime =
+          maybe (posixSecondsToUTCTime 0) (posixSecondsToUTCTime . fromInteger) regTime
+      , metadataNarBytes = if narSize == (0 :: Int) then Nothing else Just (fromIntegral narSize)
+      , metadataTrust = if ultimate then BuiltLocally else BuiltElsewhere
+      , metadataSigs = sigs
+      , metadataContentAddress = ca
       }