+++ hnix-store-json/tests/JSONSpec.hs
@@ -9,11 +9,11 @@ import Test.Hspec.QuickCheck (prop)
 
 import System.Nix.Arbitrary ()
 import System.Nix.ContentAddress (ContentAddress)
+import System.Nix.Derivation (Derivation)
 import System.Nix.DerivedPath (DerivedPath, OutputsSpec, SingleDerivedPath)
 import System.Nix.JSON ()
-import System.Nix.OutputName (OutputName)
 import System.Nix.Realisation (BuildTraceKey, Realisation)
-import System.Nix.Signature (Signature)
+import System.Nix.Signature (Signature, NamedSignature)
 import System.Nix.StorePath (StorePath, StorePathName, StorePathHashPart)
 
 roundtripsJSON
@@ -36,6 +36,8 @@ spec = do
     prop "OutputsSpec" $ roundtripsJSON @OutputsSpec
     prop "SingleDerivedPath" $ roundtripsJSON @SingleDerivedPath
     prop "DerivedPath" $ roundtripsJSON @DerivedPath
-    prop "BuildTraceKey OutputName" $ roundtripsJSON @(BuildTraceKey OutputName)
+    prop "Derivation" $ roundtripsJSON @Derivation
+    prop "BuildTraceKey" $ roundtripsJSON @BuildTraceKey
     prop "Signature" $ roundtripsJSON @Signature
+    prop "NamedSignature" $ roundtripsJSON @NamedSignature
     prop "Realisation" $ roundtripsJSON @Realisation