+++ hnix-store-remote/tests/NixSerializerSpec.hs
@@ -13,7 +13,7 @@ import Data.Time.Clock.POSIX qualified
 
 import System.Nix.Arbitrary ()
 import System.Nix.Build (BuildResult(..), BuildSuccess(..), BuildFailure(..))
-import System.Nix.Derivation (Derivation(inputDrvs))
+import System.Nix.Derivation.Traditional qualified
 import System.Nix.Store.Remote.Arbitrary ()
 import System.Nix.Store.Remote.Serializer
 import System.Nix.Store.Remote.Types.Logger (Logger(..))
@@ -74,7 +74,7 @@ spec = parallel $ do
         $ \sd ->
             roundtripS (buildResult sd (ProtoVersion 1 28))
             . (\x -> x { buildResultStatus = case buildResultStatus x of
-                          Right s -> Right s
+                          Right (BuildSuccess st _bo) -> Right (BuildSuccess st mempty)
                           Left (BuildFailure st em _nd) -> Left (BuildFailure st em False)
                        })
             . (\x -> x { buildResultTimesBuilt = 0
@@ -88,6 +88,10 @@ spec = parallel $ do
         $ \sd -> forAll (arbitrary `suchThat` ((> 28) . protoVersion_minor))
         $ \pv ->
             roundtripS (buildResult sd pv)
+            . (\x -> x { buildResultStatus = case buildResultStatus x of
+                          Right (BuildSuccess st _bo) -> Right (BuildSuccess st mempty)
+                          Left f -> Left f
+                       })
             . (\x -> x { buildResultCpuUser = Nothing
                        , buildResultCpuSystem = Nothing
                        }
@@ -114,9 +118,9 @@ spec = parallel $ do
       prop "SHA256" $ roundtripS . digest @SHA256
       prop "SHA512" $ roundtripS . digest @SHA512
 
-    prop "Derivation" $ \sd ->
-      roundtripS (derivation sd)
-      . (\drv -> drv { inputDrvs = mempty })
+    prop "Derivation" $ \sd drv ->
+      roundtripS (basicDerivation sd) $
+        System.Nix.Derivation.Traditional.withoutName drv
 
     prop "ProtoVersion" $ roundtripS @() @ProtoVersion protoVersion
 
@@ -158,7 +162,6 @@ spec = parallel $ do
 
 restrictProtoVersion :: ProtoVersion -> Some StoreRequest -> Bool
 restrictProtoVersion v (Some (BuildPaths _ _)) | v < ProtoVersion 1 30 = False
-restrictProtoVersion _ (Some (BuildDerivation _ drv _)) = inputDrvs drv == mempty
 restrictProtoVersion v (Some (QueryMissing _)) | v < ProtoVersion 1 30 = False
 restrictProtoVersion _ _ = True