+++ hnix-store-tests/src/System/Nix/Arbitrary/Realisation.hs
@@ -9,19 +9,18 @@ import System.Nix.Arbitrary.Hash ()
 import System.Nix.Arbitrary.OutputName ()
 import System.Nix.Arbitrary.Signature ()
 import System.Nix.Arbitrary.StorePath ()
-import System.Nix.Realisation (BuildTraceKey, Realisation)
+import System.Nix.Realisation (BuildTraceKey, Realisation, RealisationWithId)
 
 import Test.QuickCheck (Arbitrary(..))
-import Test.QuickCheck.Arbitrary.Generic (Arg, GenericArbitrary(..), genericArbitrary, genericShrink)
+import Test.QuickCheck.Arbitrary.Generic (GenericArbitrary(..), genericArbitrary, genericShrink)
 
-instance
-  ( Arg (BuildTraceKey outputName) outputName
-  , Arbitrary outputName
-  ) =>
-  Arbitrary (BuildTraceKey outputName)
+instance Arbitrary BuildTraceKey
   where
     arbitrary = genericArbitrary
     shrink = genericShrink
 
 deriving via GenericArbitrary Realisation
   instance Arbitrary Realisation
+
+deriving via GenericArbitrary RealisationWithId
+  instance Arbitrary RealisationWithId